Move insert_many_values, insert_values to sqlite.query_helpers
also VALUE and VALUES
