+++ qcodes/dataset/sqlite/query_helpers.py
@@ -1,7 +1,20 @@
+import itertools
 import sqlite3
+from distutils.version import LooseVersion
+from numbers import Number
 from typing import List, Any, Union, Dict, Tuple
 
-from qcodes.dataset.sqlite.connection import ConnectionPlus, atomic_transaction
+import numpy as np
+from numpy import ndarray
+
+from qcodes.dataset.sqlite.connection import ConnectionPlus, \
+    atomic_transaction, transaction, atomic
+from qcodes.dataset.sqlite.settings import SQLiteSettings
+
+
+# represent the type of  data we can/want map to sqlite column
+VALUE = Union[str, Number, List, ndarray, bool]
+VALUES = List[VALUE]
 
 
 def one(curr: sqlite3.Cursor, column: Union[int, str]) -> Any:
@@ -107,3 +120,99 @@ def update_where(conn: ConnectionPlus, table: str,
         {where_column} = ?
     """
     atomic_transaction(conn, query, *values, where_value)
+
+
+def insert_values(conn: ConnectionPlus,
+                  formatted_name: str,
+                  columns: List[str],
+                  values: VALUES,
+                  ) -> int:
+    """
+    Inserts values for the specified columns.
+    Will pad with null if not all parameters are specified.
+    NOTE this need to be committed before closing the connection.
+    """
+    _columns = ",".join(columns)
+    _values = ",".join(["?"] * len(columns))
+    query = f"""INSERT INTO "{formatted_name}"
+        ({_columns})
+    VALUES
+        ({_values})
+    """
+
+    c = atomic_transaction(conn, query, *values)
+    return c.lastrowid
+
+
+def insert_many_values(conn: ConnectionPlus,
+                       formatted_name: str,
+                       columns: List[str],
+                       values: List[VALUES],
+                       ) -> int:
+    """
+    Inserts many values for the specified columns.
+
+    Example input:
+    columns: ['xparam', 'yparam']
+    values: [[x1, y1], [x2, y2], [x3, y3]]
+
+    NOTE this need to be committed before closing the connection.
+    """
+    # We demand that all values have the same length
+    lengths = [len(val) for val in values]
+    if len(np.unique(lengths)) > 1:
+        raise ValueError('Wrong input format for values. Must specify the '
+                         'same number of values for all columns. Received'
+                         f' lengths {lengths}.')
+    no_of_rows = len(lengths)
+    no_of_columns = lengths[0]
+
+    # The TOTAL number of inserted values in one query
+    # must be less than the SQLITE_MAX_VARIABLE_NUMBER
+
+    # Version check cf.
+    # "https://stackoverflow.com/questions/9527851/sqlite-error-
+    #  too-many-terms-in-compound-select"
+    version = SQLiteSettings.settings['VERSION']
+
+    # According to the SQLite changelog, the version number
+    # to check against below
+    # ought to be 3.7.11, but that fails on Travis
+    if LooseVersion(str(version)) <= LooseVersion('3.8.2'):
+        max_var = SQLiteSettings.limits['MAX_COMPOUND_SELECT']
+    else:
+        max_var = SQLiteSettings.limits['MAX_VARIABLE_NUMBER']
+    rows_per_transaction = int(int(max_var)/no_of_columns)
+
+    _columns = ",".join(columns)
+    _values = "(" + ",".join(["?"] * len(values[0])) + ")"
+
+    a, b = divmod(no_of_rows, rows_per_transaction)
+    chunks = a*[rows_per_transaction] + [b]
+    if chunks[-1] == 0:
+        chunks.pop()
+
+    start = 0
+    stop = 0
+
+    with atomic(conn) as conn:
+        for ii, chunk in enumerate(chunks):
+            _values_x_params = ",".join([_values] * chunk)
+
+            query = f"""INSERT INTO "{formatted_name}"
+                        ({_columns})
+                        VALUES
+                        {_values_x_params}
+                     """
+            stop += chunk
+            # we need to make values a flat list from a list of list
+            flattened_values = list(
+                itertools.chain.from_iterable(values[start:stop]))
+
+            c = transaction(conn, query, *flattened_values)
+
+            if ii == 0:
+                return_value = c.lastrowid
+            start += chunk
+
+    return return_value
